// app/backend/src/interfaces/leader.interface.ts
export interface ILeader {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: number;
}

export type TeamGoals = 'homeTeamGoals' | 'awayTeamGoals';